Update zipit and creation of project 2 report structure.

+++ zipit.sh
@@ -1,21 +1,28 @@
 #!/usr/bin/env bash
-cd doc/project1
 
-./report_builder.sh
+if [ $# -eq 0 ]
+  then
+    echo "No arguments supplied, usage: ./zipit.sh (project1 | project2 | project3 | project4)"
+else
 
-cd -
+    cd "doc/$1"
 
-mkdir -p delivery/doc
+    ./report_builder.sh
 
-# all docs
-cp doc/project1/pdfs/project_1.pdf delivery/doc/project_1.pdf
+    cd -
 
-find src -name '*.py' | cpio -pdm delivery
-find res -name '*.py' -o -name '*.txt' -o -name '*.png' | cpio -pdm delivery
-find tests -name '*.py' | cpio -pdm delivery
-cp run.sh delivery/run.sh
-cp tests.sh delivery/tests.sh
+    mkdir -p delivery/doc
 
-zip -r delivery.zip delivery
+    # all docs
+    cp "doc/$1/pdfs/$1.pdf" "delivery/doc/$1.pdf"
 
-rm -r delivery
+    find src -name '*.py' -o -name '*.c' -o -name '*.h' -o -name '*.sh' | cpio -pdm delivery
+    find res -name '*.py' -o -name '*.txt' -o -name '*.png' | cpio -pdm delivery
+    find tests -name '*.py' -o -name '*.c' -o -name '*.h' | cpio -pdm delivery
+    cp run.sh delivery/run.sh
+    cp tests.sh delivery/tests.sh
+
+    zip -r delivery.zip delivery
+
+    rm -r delivery
+fi
